+++ CHANGELOG.md
@@ -23,6 +23,7 @@ We use the following categories for changes:
   configuration. Supersedes `startup.dataset.config` which accepts a string
   instead of a mapping [#1737]
 - Add alert to notify about duplicate sample/metric ingestion. [#1688]
+- Add histogram to track number samples/metadata/span sent per request [#1767]
 
 ### Changed
 - Reduced the verbosity of the logs emitted by the vacuum engine [#1715]
@@ -31,9 +32,12 @@ We use the following categories for changes:
 - In order to reduce the overall load on the system, some internal database
   metrics won't be collected as often as they used to. None of the affected 
   metrics is expected to change faster than its new collection interval [#1793]
+- Aggregate metrics at global level to avoid legend pollution in dashboards [#1800]
+- The vacuum engine now looks for compressed chunks missing stats and vacuums these too [#1804]
 
 ### Fixed
-
+- Fixing the query behind chunks_uncompressed. The new definition should
+  change the baseline value [#1794]
 
 ## [0.16.0] - 2022-10-20